requirements.txt
matplotlib==3.3.2
numpy==1.18.5
opencv-python==4.4.0.44
pdbpp==0.10.2
tensorboard==2.3.0
tensorflow-gpu==2.3.0
tensorflow-model-optimization
scipy==1.4.1
Pillow==8.0.1
gdown==3.13.0
Flask==2.1.2
Flask-Bcrypt==0.7.1
Flask-Classful==0.14.2
Flask-Cors==3.0.10
Flask-Ext==0.1
Flask-Jsonpify==1.5.0
Flask-Markdown==0.3
Flask-WTF==0.14.3
flake8==4.0.1
pytest==7.0.1
black==22.3.0
isort==5.10.1
pytest-cov==3.0.0
mypy==0.961
pytest-mock==3.7.0
pytest-flask==1.2.0
protobuf==3.20.0
chardet
